chore: Rename bpfassets to bpf
