+++ pkg/bpf/attacher_stub.go
@@ -17,7 +17,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package attacher
+package bpf
 
 type stubAttacher struct{}
