--- pkg/bpfassets/bpfassets_suite_test.go
@@ -1,13 +1,0 @@
-package bpfassets
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestBpfassets(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Bpfassets Suite")
-}
