+++ pkg/manager/manager.go
@@ -19,7 +19,7 @@ package manager
 import (
 	"time"
 
-	bpfAttacher "github.com/sustainable-computing-io/kepler/pkg/bpfassets/attacher"
+	"github.com/sustainable-computing-io/kepler/pkg/bpf"
 	"github.com/sustainable-computing-io/kepler/pkg/collector"
 	"github.com/sustainable-computing-io/kepler/pkg/config"
 	"github.com/sustainable-computing-io/kepler/pkg/kubernetes"
@@ -41,7 +41,7 @@ type CollectorManager struct {
 	Watcher *kubernetes.ObjListWatcher
 }
 
-func New(attacher bpfAttacher.Attacher) *CollectorManager {
+func New(attacher bpf.Attacher) *CollectorManager {
 	manager := &CollectorManager{}
 	manager.StatsCollector = collector.NewCollector(attacher)
 	manager.PrometheusCollector = exporter.NewPrometheusExporter()
