+++ pkg/manager/manager_test.go
@@ -6,14 +6,14 @@ package manager
 import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-	bpfAttacher "github.com/sustainable-computing-io/kepler/pkg/bpfassets/attacher"
+	"github.com/sustainable-computing-io/kepler/pkg/bpf"
 )
 
 var _ = Describe("Manager", func() {
 
 	It("Should work properly", func() {
 
-		attacher := bpfAttacher.NewMockAttacher(true)
+		attacher := bpf.NewMockAttacher(true)
 		CollectorManager := New(attacher)
 		err := CollectorManager.Start()
 		Expect(err).NotTo(HaveOccurred())
