+++ pkg/model/benchmark_test.go
@@ -19,7 +19,7 @@ import (
 	"os"
 	"testing"
 
-	bpfAttacher "github.com/sustainable-computing-io/kepler/pkg/bpfassets/attacher"
+	"github.com/sustainable-computing-io/kepler/pkg/bpf"
 	"github.com/sustainable-computing-io/kepler/pkg/collector"
 	"github.com/sustainable-computing-io/kepler/pkg/collector/stats"
 	"github.com/sustainable-computing-io/kepler/pkg/model"
@@ -33,7 +33,7 @@ func benchmarkNtesting(b *testing.B, processNumber int) {
 	// enable metrics
 	stats.SetMockedCollectorMetrics()
 	// create node node metrics
-	attacher := bpfAttacher.NewMockAttacher(true)
+	attacher := bpf.NewMockAttacher(true)
 	metricCollector := collector.NewCollector(attacher)
 
 	// create processes
